Always show the install confirm dialog, also for direct install
