+++ src-tauri/src/profile/export/commands.rs
@@ -65,10 +65,7 @@ pub fn get_pack_args(app: AppHandle) -> Result<Option<ModpackInfo>> {
 
     let hexium_exclusive = modpack::refresh_args(profile, &*app.lock_thunderstore(), game);
 
-    Ok(profile.modpack.as_ref().map(|args| ModpackInfo {
-        args: args.clone(),
-        hexium_exclusive,
-    }))
+    Ok(profile.modpack.clone().map(|args| ModpackInfo { args, hexium_exclusive }))
 }
 
 #[command]
