+++ src/lib/components/mod-list/InstallModButton.svelte
@@ -1,26 +1,27 @@
 <script lang="ts">
 	import ContextMenuContent from '$lib/components/ui/ContextMenuContent.svelte';
 	import * as api from '$lib/api';
-	import { Backend, type ContextItem, type Mod, type ModId } from '$lib/types';
+	import type { ContextItem, Mod, ModId } from '$lib/types';
 	import { shortenFileSize } from '$lib/util';
 	import Icon from '@iconify/svelte';
 	import { DropdownMenu } from 'bits-ui';
 	import DropdownArrow from '$lib/components/ui/DropdownArrow.svelte';
 	import Spinner from '../ui/Spinner.svelte';
 	import { m } from '$lib/paraglide/messages';
-	import ConfirmDialog from '$lib/components/ui/ConfirmDialog.svelte';
-	import Button from '$lib/components/ui/Button.svelte';
-	import Checkbox from '$lib/components/ui/Checkbox.svelte';
-	import Info from '$lib/components/ui/Info.svelte';
-	import Label from '$lib/components/ui/Label.svelte';
 
 	type Props = {
 		mod: Mod;
 		locked: boolean;
+		loading: boolean;
 		install: (mod: ModId) => void;
 	};
 
-	let { mod, locked, install }: Props = $props();
+	let { mod, locked, loading = $bindable(), install }: Props = $props();
+
+	let versionsOpen = $state(false);
+	let downloadSize: number | null = $state(null);
+
+	let disabled = $derived(mod.isInstalled || locked || loading);
 
 	let modId = $derived({
 		packageUuid: mod.uuid,
@@ -28,40 +29,10 @@
 		backend: mod.backend
 	});
 
-	let versionsOpen = $state(false);
-	let dialogOpen = $state(false);
-	// svelte-ignore state_referenced_locally
-	let installPayload: ModId = $state(modId);
-	let downloadSize: number | null = $state(null);
-
-	let loading = $state(false);
-	let warnNoRemind = $state(false);
-
-	let disabled = $derived(mod.isInstalled || locked || loading);
-
-	async function doInstall() {
-		if (warnNoRemind) {
-			let prefs = await api.prefs.get();
-			prefs.backendSkipConfirm = true;
-			await api.prefs.set(prefs);
-		}
-		install(installPayload);
-		loading = true;
-		dialogOpen = false;
-	}
-	async function tryInstall(modId: ModId) {
-		installPayload = modId;
-		if (mod.backend !== Backend.Thunderstore && !(await api.prefs.get()).backendSkipConfirm) {
-			dialogOpen = true;
-		} else {
-			doInstall();
-		}
-	}
-
 	let contextItems: ContextItem[] = $derived(
 		mod.versions.map((version) => ({
 			label: version.name,
-			onclick: async () => tryInstall({
+			onclick: () => install({
 				packageUuid: mod.uuid,
 				versionUuid: version.uuid,
 				backend: mod.backend
@@ -70,7 +41,6 @@
 	);
 
 	$effect(() => {
-		loading = false;
 		api.profile.install.getDownloadSize(modId).then((size) => (downloadSize = size));
 	});
 </script>
@@ -78,7 +48,7 @@
 <div class="mt-2 flex text-lg text-white">
 	<button
 		class="enabled:bg-accent-700 enabled:hover:bg-accent-600 disabled:bg-primary-700 disabled:text-primary-300 flex grow items-center justify-center gap-2 rounded-l-lg py-2 font-semibold disabled:cursor-not-allowed"
-		onclick={() => tryInstall(modId)}
+		onclick={() => install(modId)}
 		{disabled}
 	>
 		{#if locked}
@@ -112,19 +82,3 @@
 		/>
 	</DropdownMenu.Root>
 </div>
-
-<ConfirmDialog title={m.otherServer_warn_title()} bind:open={dialogOpen}>
-	{m.otherServer_warn_content()}
-	<div class="my-5 flex items-center">
-		<Checkbox id="neverwarninstall" bind:checked={warnNoRemind} />
-		<label class="ml-3" for="neverwarninstall">
-			{m.otherServer_warn_noremind()}
-		</label>
-	</div>
-
-	{#snippet buttons()}
-		<Button color="accent" icon="mdi:download" onclick={doInstall}>
-			{m.installModButton_button_install()}
-		</Button>
-	{/snippet}
-</ConfirmDialog>
